+++ flatlib/const.py
@@ -177,6 +177,7 @@ ASC = 'Asc'
 DESC = 'Desc'
 MC = 'MC'
 IC = 'IC'
+VERTEX = 'Vertex'
 
 # === Fixed Stars === */
 
@@ -295,7 +296,7 @@ LIST_HOUSES = [
 ]
 
 LIST_ANGLES = [
-    ASC, MC, DESC, IC
+    ASC, MC, DESC, IC, VERTEX
 ]
 
 LIST_ASTEROIDS = [